fix：官网邮件的链接问题

+++ ses-app/ses-web-ros/src/main/java/com/redescooter/ses/web/ros/vo/sales/SalesOrderDetailsResult.java
@@ -96,12 +96,12 @@ public class SalesOrderDetailsResult extends GeneralResult {
     private Double prepaidDeposit;
 
     @ApiModelProperty("规格类型id")
-    private Double specificatTypeId;
+    private Long specificatTypeId;
 
     @ApiModelProperty("客户咨询管理id")
-    private Double customerInquiryId;
+    private Long customerInquiryId;
 
     @ApiModelProperty("颜色id")
-    private Double colorId;
+    private Long colorId;
 
 }
